pyproject.toml
[project]
name = "pokebot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.7.1",
    "fastapi[standard]>=0.136.0",
    "langchain-core>=1.3.0",
    "langchain-google-genai>=4.2.2",
    "langchain-openai>=1.1.16",
    "langgraph>=1.1.0",
    "pillow>=12.2.0",
    "python-dotenv>=1.2.2",
]
